Implementado giro de pantalla. Implementada base de datos (aún no se usa). Empezada pantalla de login. Cambiada nomenclatura de oxígeno (1.000 => 1K).

+++ app/src/main/java/com/example/entrega1/MejorasSegundo.java
@@ -30,9 +30,9 @@ public class MejorasSegundo extends Fragment {
 
     private String nombres[] = {"Cohete", "Nave", "Ovni", "ISS", "Algo", "Algo", "Algo", "Algo", "Algo", "Algo", "Algo", "Algo"};
 
-    private Integer cantidades[] = {1, 5, 20, 80, 100, 100, 100, 100, 100, 100, 100, 100, 100};
+    private float cantidades[] = {1, 5, 20, 80, 100, 100, 100, 100, 100, 100, 100, 100, 100};
 
-    private Integer precios[] = {10, 50, 200, 800, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000};
+    private float precios[] = {10, 50, 200, 800, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000};
 
     public MejorasSegundo() {
         // Required empty public constructor
